+++ doc.go
@@ -42,8 +42,8 @@ operation being performed as part of a traced workflow. It its simplest form:
 A Tracer is unique to the instrumentation and is used to create Spans.
 Instrumentation should be designed to accept a TracerProvider from which it
 can create its own unique Tracer. Alternatively, the registered global
-TracerProvider from the go.opentelemetry.io/otel/api/global package can be
-used as a default.
+TracerProvider from the go.opentelemetry.io/otel/global package can be used as
+a default.
 
 	const (
 		name    = "instrumentation/package/name"
