+++ example/jaeger/main.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"log"
 
-	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/global"
 	"go.opentelemetry.io/otel/label"
 
 	"go.opentelemetry.io/otel/exporters/trace/jaeger"
