+++ global/internal/trace_test.go
@@ -20,8 +20,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"go.opentelemetry.io/otel/api/global"
-	"go.opentelemetry.io/otel/api/global/internal"
+	"go.opentelemetry.io/otel/global"
+	"go.opentelemetry.io/otel/global/internal"
 	"go.opentelemetry.io/otel/oteltest"
 )
