+++ internal/metric/async.go
@@ -21,7 +21,7 @@ import (
 	"sync"
 
 	api "go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/global"
 	"go.opentelemetry.io/otel/label"
 )
