+++ propagators/trace_context_example_test.go
@@ -15,7 +15,7 @@
 package propagators_test
 
 import (
-	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/global"
 	"go.opentelemetry.io/otel/propagators"
 )
