+++ sdk/metric/sdk.go
@@ -23,7 +23,7 @@ import (
 
 	"go.opentelemetry.io/otel"
 	api "go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/global"
 	internal "go.opentelemetry.io/otel/internal/metric"
 	"go.opentelemetry.io/otel/label"
 	export "go.opentelemetry.io/otel/sdk/export/metric"
